requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
bcrypt
python-jose
python-decouple
python-multipart